Change root path to join with post to game.
Changes the root path to be the join page, and updates this to post to
the game endpoint.

The game endpoint now creates a player and places the UUID for the
player on a data attribute for the game. This can then hopefully be
extracted and used to instantiate a player locally that matches the one
stored in redis.