Hardcode webpack dev server port.
Sometimes the webpack dev server process isn't being killed when running
development environment.

This leads to a new server being spawned on next boot but on a different
port. Hopefully this hack will at least lead to an error when binding to
to 8081 so that developer knows they need to manually kill node process.