Refactors subscription to channel in to Player.
Player constructor takes a connection and channel and subscribes the
player to the channel when calling `Player#join`.

This includes pushing the player JSON into redis. so that it can be
later broadcast as a list of connected players.

The Player name isn't yet correctly stored in Redis, will need to be
sent to the server after connection.