Remove player from lobby and broadcast on disconnect.
Refactors some the connection events out of the Subscription class and
back up in to the main Socket block, this makes it a bit clearer to see
the lifecycle of the connection.

When a connection disconnects the player is removed from the lobby and
updated lobby is broadcast to all connections in the channel.