Add settings objects to build socket/web URLs.
The ruby Settings object builds the correct socket and web URLs based on
the environment and BASE_HOST ENV values.

The socket URL is set as a data attribute and then read in by the JS
Settings object and passed to Sockets.

This make the assumption that the app will be using HTTP/WS in
development.